Plateau implemente dans le modele plateau
